+++ user-center-service/src/main/java/com/cfx/usercenter/entity/Menu.java
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
  * </p>
  *
  * @author zhangziyao
- * @since 2023-05-06
+ * @since 2023-05-08
  */
 @Data
 @EqualsAndHashCode(callSuper = false)
@@ -28,12 +28,12 @@ public class Menu implements Serializable {
      * 资源ID
      */
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 系统id
      */
-    private Integer appId;
+    private Long appId;
 
     /**
      * 资源名称
@@ -58,7 +58,7 @@ public class Menu implements Serializable {
     /**
      * 上级资源ID
      */
-    private Integer parentId;
+    private Long parentId;
 
     /**
      * 资源级别
